Update part_3_1.py part_3_2.py solution to problem 3.2, fixing bug in problem 3.1.

+++ part_3_1.py
@@ -56,7 +56,8 @@ def problem_3_1():
         os.path.dirname(__file__),
         result_filename
     ))
-    os.remove(result_abspath)
+    if os.path.exists(result_abspath):
+        os.remove(result_abspath)
 
     _fp = open(result_abspath, 'w')
     writer = csv.writer(_fp)
